requirements.txt
asgiref==3.6.0
Django==3.2.16
djangorestframework==3.14.0
gunicorn==20.1.0
pytz==2022.7
sqlparse==0.4.3
typing-extensions==4.4.0
uuid==1.30
